RS/datamodel: Refactoring, split datamodel into different files.
