+++ server/prisma-rs/libs/datamodel/src/dml/comment.rs
@@ -1,0 +1,6 @@
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct Comment {
+    pub text: String,
+    pub is_error: bool,
+}
